Refactor generate_exec_scripts.py: Modularize code and improve readability
- Removed extensive inline functions and replaced them with imports from new modules.
- Introduced a structured approach by separating concerns into different files:
  - `metadata.py` for metadata generation.
  - `paths.py` for path-related utilities.
  - `script_generation.py` for script generation logic.
  - `validation.py` for JSON schema validation.
- Updated the main execution block to utilize the new modular structure.
- Improved argument parsing and handling of workspace directory.
- Enhanced error handling and directory management during script generation.
