+++ manager_scripts/generate_exec/__init__.py
@@ -1,0 +1,1 @@
+"""Submodules for topology validation and execution-script generation."""
